# api_rede_social/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Protótipo de Rede Social",
        "description": "API protótipo de rede social.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/usuarios": {
            "get": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Listar todos os usuários.",
                "operationId": "3ec8bd4e29c5f17719542f5bd741e3f8",
                "responses": {
                    "200": {
                        "description": "Lista de usuários retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "ID do usuário",
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "description": "Nome do usuário",
                                                "type": "string",
                                                "example": "João"
                                            },
                                            "email": {
                                                "description": "Email do usuário",
                                                "type": "string",
                                                "example": "joao@exemplo.com"
                                            },
                                            "password": {
                                                "description": "ID do usuário",
                                                "type": "string",
                                                "example": "xxsenha123"
                                            },
                                            "created_at": {
                                                "description": "momento em que o registro foi criado no banco de dados.",
                                                "type": "string",
                                                "example": "2025-01-02T20:00:39.000000Z"
                                            },
                                            "updated_at": {
                                                "description": "momento em que o registro foi atualizado pela última vez",
                                                "type": "string",
                                                "example": "2025-01-02T20:00:39.000000Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Cadastrar um novo usuário.",
                "operationId": "fdccd67c0494d38b46d588d3a1a0880c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Nome do usuário",
                                        "type": "string",
                                        "example": "João"
                                    },
                                    "email": {
                                        "description": "Email do usuário",
                                        "type": "string",
                                        "example": "joao@exemplo.com"
                                    },
                                    "password": {
                                        "description": "Senha do usuário",
                                        "type": "string",
                                        "example": "senha123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID do usuário",
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "description": "Nome do usuário",
                                            "type": "string",
                                            "example": "João"
                                        },
                                        "email": {
                                            "description": "Email do usuário",
                                            "type": "string",
                                            "example": "joao@exemplo.com"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos ou faltando parâmetros",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erro ao criar o usuário."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/usuarios/{id}": {
            "get": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Exibir os dados de um usuário específico.",
                "operationId": "12a45471173c6151fc4d28ae3f3cc8c2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuário encontrado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID do usuário",
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "description": "Nome do usuário",
                                            "type": "string",
                                            "example": "João"
                                        },
                                        "email": {
                                            "description": "Email do usuário",
                                            "type": "string",
                                            "example": "joao@exemplo.com"
                                        },
                                        "password": {
                                            "description": "ID do usuário",
                                            "type": "string",
                                            "example": "xxsenha123"
                                        },
                                        "created_at": {
                                            "description": "momento em que o registro foi criado no banco de dados.",
                                            "type": "string",
                                            "example": "2025-01-02T20:00:39.000000Z"
                                        },
                                        "updated_at": {
                                            "description": "momento em que o registro foi atualizado pela última vez",
                                            "type": "string",
                                            "example": "2025-01-02T20:00:39.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuário não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Usuário não encontrado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Atualizar os dados de um usuário específico.",
                "operationId": "356f4dacfa0a376c973f916ac401f8ef",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Nome do usuário",
                                        "type": "string",
                                        "example": "João"
                                    },
                                    "email": {
                                        "description": "Email do usuário",
                                        "type": "string",
                                        "example": "joao@exemplo.com"
                                    },
                                    "password": {
                                        "description": "Senha do usuário",
                                        "type": "string",
                                        "example": "senha123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuário atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID do usuário",
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "description": "Nome do usuário",
                                            "type": "string",
                                            "example": "João"
                                        },
                                        "email": {
                                            "description": "Email do usuário",
                                            "type": "string",
                                            "example": "joao@exemplo.com"
                                        },
                                        "password": {
                                            "description": "ID do usuário",
                                            "type": "string",
                                            "example": "xxsenha123"
                                        },
                                        "created_at": {
                                            "description": "momento em que o registro foi criado no banco de dados.",
                                            "type": "string",
                                            "example": "2025-01-02T20:00:39.000000Z"
                                        },
                                        "updated_at": {
                                            "description": "momento em que o registro foi atualizado pela última vez",
                                            "type": "string",
                                            "example": "2025-01-02T20:00:39.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos ou faltando parâmetros",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Erro ao atualizar o usuário."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "ID inválido. Usuário não encontrado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "O usuário não existe."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Deletar um usuário específico.",
                "operationId": "067bf89c772e70e9b85f1fbe643e7cc8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Usuário deletado com sucesso"
                    },
                    "404": {
                        "description": "Usuário não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Usuário não encontrado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}